+++ bootstrap/src/main/java/onlydust/com/marketplace/api/bootstrap/configuration/DomainConfiguration.java
@@ -6,22 +6,23 @@ import onlydust.com.marketplace.accounting.domain.port.in.AccountingFacadePort;
 import onlydust.com.marketplace.accounting.domain.port.in.CurrencyFacadePort;
 import onlydust.com.marketplace.accounting.domain.port.out.*;
 import onlydust.com.marketplace.accounting.domain.service.CurrencyService;
-import onlydust.com.marketplace.project.domain.gateway.DateProvider;
-import onlydust.com.marketplace.project.domain.job.IndexerApiOutboxConsumer;
-import onlydust.com.marketplace.project.domain.job.OutboxConsumer;
-import onlydust.com.marketplace.project.domain.job.OutboxConsumerJob;
-import onlydust.com.marketplace.project.domain.job.WebhookNotificationOutboxConsumer;
-import onlydust.com.marketplace.project.domain.observer.ContributionObserver;
-import onlydust.com.marketplace.project.domain.observer.ProjectObserver;
-import onlydust.com.marketplace.project.domain.observer.UserObserver;
-import onlydust.com.marketplace.project.domain.service.*;
 import onlydust.com.marketplace.api.infrastructure.accounting.AccountingServiceAdapter;
 import onlydust.com.marketplace.api.postgres.adapter.*;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.UserVerificationEventEntity;
 import onlydust.com.marketplace.api.sumsub.webhook.adapter.mapper.SumsubMapper;
+import onlydust.com.marketplace.kernel.jobs.OutboxConsumerJob;
 import onlydust.com.marketplace.kernel.port.output.ImageStoragePort;
+import onlydust.com.marketplace.kernel.port.output.OutboxConsumer;
+import onlydust.com.marketplace.kernel.port.output.OutboxPort;
+import onlydust.com.marketplace.project.domain.gateway.DateProvider;
+import onlydust.com.marketplace.project.domain.job.IndexerApiOutboxConsumer;
+import onlydust.com.marketplace.project.domain.job.WebhookNotificationOutboxConsumer;
+import onlydust.com.marketplace.project.domain.observer.ContributionObserver;
+import onlydust.com.marketplace.project.domain.observer.ProjectObserver;
+import onlydust.com.marketplace.project.domain.observer.UserObserver;
 import onlydust.com.marketplace.project.domain.port.input.*;
 import onlydust.com.marketplace.project.domain.port.output.*;
+import onlydust.com.marketplace.project.domain.service.*;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.retry.annotation.EnableRetry;
