+++ infrastructure/postgres-adapter/src/main/java/onlydust/com/marketplace/api/postgres/adapter/repository/NotificationEventRepository.java
@@ -1,6 +1,6 @@
 package onlydust.com.marketplace.api.postgres.adapter.repository;
 
-import onlydust.com.marketplace.project.domain.model.notification.Event;
+import onlydust.com.marketplace.kernel.model.Event;
 import onlydust.com.marketplace.api.postgres.adapter.OutboxRepository;
 import onlydust.com.marketplace.api.postgres.adapter.entity.write.NotificationEventEntity;
