+++ infrastructure/posthog-api-client-adapter/src/main/java/onlydust/com/marketplace/api/posthog/adapters/PosthogApiClientAdapter.java
@@ -3,7 +3,7 @@ package onlydust.com.marketplace.api.posthog.adapters;
 import io.netty.handler.codec.http.HttpMethod;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import onlydust.com.marketplace.project.domain.model.notification.Event;
+import onlydust.com.marketplace.kernel.model.Event;
 import onlydust.com.marketplace.project.domain.model.notification.UserSignedUp;
 import onlydust.com.marketplace.project.domain.port.output.WebhookPort;
 import onlydust.com.marketplace.api.posthog.client.PosthogHttpClient;
