+++ kernel/src/main/java/onlydust/com/marketplace/kernel/jobs/OutboxSkippingException.java
@@ -1,4 +1,4 @@
-package onlydust.com.marketplace.project.domain.job;
+package onlydust.com.marketplace.kernel.jobs;
 
 public class OutboxSkippingException extends RuntimeException {
     public OutboxSkippingException(final String message) {
