+++ kernel/src/main/java/onlydust/com/marketplace/kernel/model/Event.java
@@ -1,4 +1,4 @@
-package onlydust.com.marketplace.project.domain.model.notification;
+package onlydust.com.marketplace.kernel.model;
 
 import lombok.EqualsAndHashCode;
