+++ kernel/src/main/java/onlydust/com/marketplace/kernel/port/output/OutboxConsumer.java
@@ -1,0 +1,7 @@
+package onlydust.com.marketplace.kernel.port.output;
+
+import onlydust.com.marketplace.kernel.model.Event;
+
+public interface OutboxConsumer {
+    void process(Event event);
+}
