+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/job/IndexerApiOutboxConsumer.java
@@ -2,7 +2,8 @@ package onlydust.com.marketplace.project.domain.job;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import onlydust.com.marketplace.project.domain.model.notification.Event;
+import onlydust.com.marketplace.kernel.model.Event;
+import onlydust.com.marketplace.kernel.port.output.OutboxConsumer;
 import onlydust.com.marketplace.project.domain.model.notification.ProjectLinkedReposChanged;
 import onlydust.com.marketplace.project.domain.model.notification.UserSignedUp;
 import onlydust.com.marketplace.project.domain.port.output.IndexerPort;
