--- project-domain/src/main/java/onlydust/com/marketplace/project/domain/job/OutboxConsumer.java
@@ -1,7 +1,0 @@
-package onlydust.com.marketplace.project.domain.job;
-
-import onlydust.com.marketplace.project.domain.model.notification.Event;
-
-public interface OutboxConsumer {
-    void process(Event event);
-}
