+++ project-domain/src/main/java/onlydust/com/marketplace/project/domain/port/input/UserVerificationFacadePort.java
@@ -1,6 +1,6 @@
 package onlydust.com.marketplace.project.domain.port.input;
 
-import onlydust.com.marketplace.project.domain.model.notification.Event;
+import onlydust.com.marketplace.kernel.model.Event;
 
 public interface UserVerificationFacadePort {
